quota querying and tree accounting (#1405)

+++ pkg/storage/fs/s3/s3.go
@@ -262,7 +262,7 @@ func (fs *s3FS) UpdateGrant(ctx context.Context, ref *provider.Reference, g *pro
 	return errtypes.NotSupported("s3: operation not supported")
 }
 
-func (fs *s3FS) GetQuota(ctx context.Context) (int, int, error) {
+func (fs *s3FS) GetQuota(ctx context.Context) (uint64, uint64, error) {
 	return 0, 0, nil
 }
